requirements.txt
flask==1.1.2
numpy==1.18.4
scipy==1.4.1
pandas == 1.0.3
Pillow == 7.1.2
flasgger == 0.9.4
#zipfile
XlsxWriter==1.2.8
stemming==1.0.1
scikit-learn==0.22.2
tensorflow==1.13.1
tensorboard==1.13.1
Keras==2.2.4
tensorboard==1.13.1
gunicorn==20.0.4
matplotlib==3.2.1